Testing reorganization, update contributing.md
1. rewrote contributing.md
2. separated tests using nodeunit tool and chai-passport-strategy tool
3. added detailed instructions on test in contributing.md
4. add some simple instructions on test in README.md
